Move obsolete version to v1 package
