+++ restclient/v1/logger.go
@@ -1,4 +1,4 @@
-package restclient
+package v1
 
 import (
 	"fmt"
