+++ restclient/v1/request_test.go
@@ -1,4 +1,4 @@
-package restclient
+package v1
 
 import (
 	"bytes"
